// frontend/features/dashboard/types/ProposalBackendData.type.ts
export interface ProposalBackendDataType {
    id: number;
    brandName: string;
    ownerName: string;
    email: string; 
    phoneNumber: string;
    city: string;
    dangerCat: string;
    employeeCount: string;
    whenStart: string;
    osgbService: boolean;
    workplacePhysician: boolean;
    explosionProtectionDocument: boolean;
    ohsSpecialist: boolean;
    ambientMeasurements: boolean;
    mobileHealthVehicle: boolean;
    workplaceNurse: boolean;
    basicOhsTraining: boolean;
    workingAtHeightTraining: boolean;
    createdAt: string;
}